♻️ refactor: download script to avoid overwrite existing data

+++ src/cartola/download_data.py
@@ -1,5 +1,5 @@
 from datetime import date
-from os.path import join
+from os.path import exists, join
 
 import pandas as pd
 import requests
@@ -22,10 +22,12 @@ year = date.today().year
 file = join("data", "01_raw", str(year), f"rodada-{rodada}.csv")
 print(file)
 
-cols_scouts = sorted([col for col in df_merge.columns if col.isupper()])
-cols_atleta = sorted(set(df_merge.columns) - set(cols_scouts))
+cols_scouts = [col for col in df_merge.columns if col.isupper()]
+cols_atleta = list(set(df_merge.columns) - set(cols_scouts))
 
 df_merge = df_merge.loc[:, cols_atleta + cols_scouts]
 df_merge.sort_values(by="atletas.atleta_id", inplace=True, ignore_index=True)
-df_merge.to_csv(file)
-# json.dump(data_json, open(file.replace(".csv", ".json"), "w", encoding="utf-8"), indent=2, ensure_ascii=False)
+
+if not exists(file):
+    df_merge.to_csv(file)
+    # json.dump(data_json, open(file.replace(".csv", ".json"), "w", encoding="utf-8"), indent=2, ensure_ascii=False)
